feat: added recipe add/edit screen

+++ app/src/main/java/com/ihita/wholetthemcook/viewmodel/RecipeInfoViewModel.kt
@@ -8,7 +8,7 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
 
-class RecipeInfoViewModel(private val recipeId: Int) : ViewModel() {
+class RecipeInfoViewModel(private val recipeId: Long) : ViewModel() {
 
     private val _recipe = MutableStateFlow<Recipe?>(null)
     val recipe: StateFlow<Recipe?> = _recipe
