+++ app/src/main/java/com/ihita/wholetthemcook/viewmodel/RecipeInfoViewModelFactory.kt
@@ -3,7 +3,7 @@ package com.ihita.wholetthemcook.viewmodel
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 
-class RecipeInfoViewModelFactory(private val recipeId: Int) : ViewModelProvider.Factory {
+class RecipeInfoViewModelFactory(private val recipeId: Long) : ViewModelProvider.Factory {
 
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
         if (modelClass.isAssignableFrom(RecipeInfoViewModel::class.java)) {
